pyproject.toml
[project]
name = "lecl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "langchain[google-genai]>=0.3.26",
    "python-dotenv>=1.0.0",
]
